feat: Creación factoría para admin #426
