+++ src/main/java/com/skillstack/devhub/factorymethod/AdminUserFactory.java
@@ -1,14 +1,13 @@
 package com.skillstack.devhub.factorymethod;
 
 import com.skillstack.devhub.model.AbstractUser;
+import com.skillstack.devhub.model.AdminUser;
 import com.skillstack.devhub.model.Role;
-import com.skillstack.devhub.model.User;
-import org.springframework.stereotype.Service;
 
-@Service
-public class DefaultUserFactory implements UserFactory{
+public class AdminUserFactory implements UserFactory{
+
     @Override
     public AbstractUser createUser(String firstName, String lastName, String username, String email, String password, String phone, Role role) {
-        return new User(firstName, lastName, username, email, password, phone, role, 0);
+        return new AdminUser(firstName, lastName, username, email, password, phone, role);
     }
 }
