+++ src/main/java/com/skillstack/devhub/factorymethod/UserFactory.java
@@ -4,5 +4,5 @@ import com.skillstack.devhub.model.AbstractUser;
 import com.skillstack.devhub.model.Role;
 
 public interface UserFactory {
-    AbstractUser createUser (String firstName, String lastName, String username, String email, String password, String phone, Role role, int totalScore);
+    AbstractUser createUser (String firstName, String lastName, String username, String email, String password, String phone, Role role);
 }
