+++ src/main/java/com/skillstack/devhub/model/AbstractUser.java
@@ -11,9 +11,8 @@ public abstract class AbstractUser {
     protected String email;
     protected String password;
     protected String phone;
-    protected List<String> preferences;
     protected Role role;
-    protected int totalScore;
+
 
     public String getFirstName() {
         return firstName;
@@ -63,14 +62,6 @@ public abstract class AbstractUser {
         this.phone = phone;
     }
 
-    public List<String> getPreferences() {
-        return preferences;
-    }
-
-    public void setPreferences(List<String> preferences) {
-        this.preferences = preferences;
-    }
-
     public Role getRole() {
         return role;
     }
@@ -79,11 +70,4 @@ public abstract class AbstractUser {
         this.role = role;
     }
 
-    public int getTotalScore() {
-        return totalScore;
-    }
-
-    public void setTotalScore(int totalScore) {
-        this.totalScore = totalScore;
-    }
 }
