+++ src/main/java/com/skillstack/devhub/service/AuthenticationService.java
@@ -68,7 +68,7 @@ public class AuthenticationService {
         }
 
         User u = (User) defaultUserFactory.createUser(user.getFirstName(), user.getLastName(), user.getUsername(), user.getEmail(),
-                user.getPassword(), user.getPhone(), Role.USER, 0);
+                user.getPassword(), user.getPhone(), Role.USER);
         u.setPassword(passwordEncoder.encode(user.getPassword()));
 
         if(userRepository.count()==0){
